download xdebug if not exist

+++ examples/base-php-app/docker-app-entrypoint.sh
@@ -8,7 +8,9 @@ fi
 
 if [ ! -z "$XDEBUG_CONFIG" ]; then
     XDEBUG_SO_PATH="/usr/local/lib/php/extensions/xdebug.so"
-    wget -O "$XDEBUG_SO_PATH" https://github.com/danilin-em/xdebug-builder/releases/download/v2.9.0/xdebug.so
+    if [ ! -f "$XDEBUG_SO_PATH" ]; then
+        wget -O "$XDEBUG_SO_PATH" https://github.com/danilin-em/xdebug-builder/releases/download/v2.9.0/xdebug.so
+    fi
     set -- "$@" -d "zend_extension=$XDEBUG_SO_PATH"
 fi
